finished collecting recon data

+++ src/main/java/org/yesworkflow/recon/UriVariable.java
@@ -2,11 +2,11 @@ package org.yesworkflow.recon;
 
 public class UriVariable
 {
-    public Integer id;
+    public Long id;
     public Integer port;
     public String name;
 
-    public UriVariable(Integer id, Integer port, String name)
+    public UriVariable(Long id, Integer port, String name)
     {
         this.id = id;
         this.port = port;
